cleaning up & layout update
