+++ cec/testpost.md
@@ -4,8 +4,4 @@ author: 原批
 date: 2023-5-??
 ---
 
-<script>
-  import Weird from "$lib/post/weird.svelte";
-</script>
-
 <Weird />
