+++ package.json
@@ -28,12 +28,11 @@
     "eslint-plugin-svelte3": "^4.0.0",
     "fast-glob": "^3.3.0",
     "highlight.js": "^11.8.0",
-    "mdsvex": "^0.10.6",
-    "minimatch": "^9.0.1",
     "postcss": "^8.4.23",
     "postcss-load-config": "^4.0.1",
     "prettier": "^2.8.0",
     "prettier-plugin-svelte": "^2.8.1",
+    "rehype-raw": "^6.1.1",
     "svelte": "^3.54.0",
     "svelte-check": "^3.0.1",
     "svelte-exmarkdown": "^2.1.0",
