+++ src/app.postcss
@@ -14,4 +14,12 @@ body {
   @apply code;
 }
 
+.prose a.permalink {
+  @apply opacity-0 pl-2 transition-all text-sm no-underline leading-4;
+}
+
+.prose :is(h1, h2, h3, h4, h5, h6):hover a.permalink {
+  @apply opacity-100;
+}
+
 /* vim: set ft=css: */
