+++ src/lib/components/weird.svelte
